change view sb_albums

+++ db/couchdb/views/sb_albums_all/map.js
@@ -1,5 +1,5 @@
 function(doc) {
   if(doc['couchrest-type'] == 'SbAlbum') {
-    emit(parseInt(doc.book_num), null);
+    emit(parseFloat(doc.book_num), null);
   }
 }
